requirements.txt
anyio==4.12.0
asgiref==3.11.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
filelock==3.20.2
flask==3.1.2
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.2.3
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pyjwt==2.10.1
python-dotenv==1.2.1
python3-openid==3.2.0
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
sympy==1.14.0
tqdm==4.67.1
typer-slim==0.21.0
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
werkzeug==3.1.4
whitenoise==6.11.0
